smoothing and handyness
made a small park in the city and smooth some un-even tiles even. in pillars spawn points for pillarmen and vimpires now all over the map and made the medbay in the fort a little smaller.  add in a small commandend center and add in a specail building for the tank in the front of both base's where all the tools you need and ammo and fuel is.